uses predefines symbolic information for run test
The run test requires a symbolier that is capable of resolving all
external symbols. Neither IDA nor objdump are present on the testing
machine. We can install the later, however it is unable to handle arm
binaries correctly (few symbols, such as malloc and strlen are still
missing). Since the task is not to check the symbolizer, but to check
the run pass we will just upload the symbolic information.