Primus (#15)
* changes the testing binary, drops gnueabihf

The new testing program is a little bit more complex, it includes some
string parsing, calling to different library functions etc. It is
designed to test the Primus emulator, Beagle and other new stuff. I will
push the tests a little bit later.

I also decided to drop support for gnueabihf abi, because byteweight
doesn't support it, and we have no time, to get it into a ready
condition (as well as no desire). Once we will have either, we will put
it back.

I also added few tests, for the missing x86 platform.

* adds tests for primus
Also updates the test to a correct (I hope) version of abi.

* updates tests to the new interface

Also the windows binaries were for some reason updated, after I
rebuilt the binaries from the sources, though the sources are still
the same. Maybe coff files contain date or something like this.

* uses predefines symbolic information for run test

The run test requires a symbolier that is capable of resolving all
external symbols. Neither IDA nor objdump are present on the testing
machine. We can install the later, however it is unable to handle arm
binaries correctly (few symbols, such as malloc and strlen are still
missing). Since the task is not to check the symbolizer, but to check
the run pass we will just upload the symbolic information.