fixes tests in order to add llvm 6.0 support
We got few fails when running tests with bap build with llvm 6.0.

First of all we have to delete 2 cases from `x86 mov` tests:
```
1. 0x8c,0x20
2. 0x8e,0x20
```
Those tests check wrong behaviour, that was fixed in llvm 6.0.
What we had in previous llvm versions:

```
$ echo "0x8c,0x20" | llvm-mc-4.0 --arch=x86 --show-inst --disassemble
	.text
	movl	%fs, (%eax)             # <MCInst #1676 MOV32ms
                                        # <MCOperand Reg:19>
                                        # <MCOperand Imm:1>
                                        # <MCOperand Reg:0>
                                        # <MCOperand Imm:0>
                                        # <MCOperand Reg:0>
                                        # <MCOperand Reg:32>>
```
And what do we have when using llvm 6.0:
```
$ echo "0x8c,0x20" | llvm-mc --arch=x86 --show-inst --disassemble
	.text
	movw	%fs, (%eax)             # <MCInst #1761 MOV16ms
                                        #  <MCOperand Reg:19>
                                        #  <MCOperand Imm:1>
                                        #  <MCOperand Reg:0>
                                        #  <MCOperand Imm:0>
                                        #  <MCOperand Reg:0>
                                        #  <MCOperand Reg:32>>
```
Assembler strings are different for the same opcodes, and it is
the reason why do we fail.

According to a manual:
```
MOV reg16/32/64/mem16,segReg     8C /r
Move the contents of a segment register to a 16-bit, 32-
bit, or 64-bit destination register or to a 16-bit memory
operand.

MOV segReg, reg/mem16            8E /r
Move the contents of a 16-bit register or memory
operand to a segment register.
```
Instructions in question deal with memory, so only 16 bits affected.
And therefore, `movw` must be in assembler string, but not `movl`.

Also, we got a less interesting fail in mips `mov` test, because
there were an extra space in assembler string in previos llvm
versions.