adds close;wait to all tests
apparently and quite surprisingly dejagnu doesn't do any resource
management for us, and as soon as we trigger `pass` or `fail` it
jumps to the next spawn leaving the currently run process open.

I'm not sure is there a better way, but so far I've added `close;wait`
before each `fail` or `pass` (where spawn was used). Probably, we
should put this in the library.