adds symbols for the raw-source test case
Since 1001 we no longer materialize functions if they are not marked
as so in the knowledge base, so we have to specify that the first
address starts a function.

The easiest way to do it right now is using the read-symbols plugin,
but we should provide something more coveninent, that won't require
a user to write a file with more information than needed.