changes bap-veri installation on Travis
this PR changes the way we install bap-veri on travis:
now opam is used to install dependencies only, and the
package itself is installed with make && make install.

the reason is that bap-veri depends both from bap and monads
library, and that complicate build on travis, since we install
all bap libraries as a sinlge package named bap.