adds the symbolic executor tests
so far we pass 7 tests (see src/symexec*.c) on 3 architectures, vid,
arm, powerpc, and x86-64, but fail on mips and mingw. In case of
mips, we just don't have any names (despite having objdump and -g),
so it is irrelevant to the symexec. On mingw, didn't even look yet,
most likely some unsupported ABI.