expects fail in macho loader
As it's visible from our nightly tests, the macho
loader fails on macOS with `ko/powerpc-xnu-kext`  file.
The reason is in the small hack in ogre loader,
thats work for ELF files, but does not for the others ones.

When we deal with relocatable files, we want to see the same addresses,
as they are in the objdump.  And we subtract an address of the first
section from the base to reach it.

But the same thing doesn't work for other files, so we have to
expect fails in the macho loaders test for some time, while this bug
won't be fixed.